move common parsing
