+++ container-core/src/main/java/com/yahoo/container/logging/LevelsModSpec.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.log.internal;
+package com.yahoo.container.logging;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
